@@ -3,27 +3,22 @@ description: >-
   A Flutter package that implements Material Design Feature discovery
   to show a description of specific features to new users.
   See https://tinyurl.com/FeatureDiscovery
-version: 0.14.1
+version: 0.14.2
 homepage: https://github.com/ayalma/feature_discovery
 
 environment:
   # This should match the Flutter stable SDK constraint.
   # See https://github.com/flutter/flutter/blob/stable/packages/flutter/pubspec.yaml.
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.3.0-0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  provider: ^6.0.0
-  shared_preferences: ^2.0.3
+  provider: ^6.1.2
+  shared_preferences: ^2.3.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-  pedantic: 1.9.0
-
-dependency_overrides:
-  # TODO(creativecreatorormaybenot): Remove when Pedantic is updated in flutter_test.
-  pedantic: 1.9.0
+  lints: any
