@@ -1,6 +1,5 @@
 import 'package:feature_discovery/feature_discovery.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'widgets.dart';
@@ -202,8 +201,7 @@ void main() {
         // The Container that makes the content of the feature overlay of the test widget has a static
         // height of 9e3, which ensures that the content definitely covers the 4e3 surface size height
         // if OverflowMode.clipContent is not enabled.
-        await (TestWidgetsFlutterBinding.ensureInitialized()
-                as TestWidgetsFlutterBinding)
+        await (TestWidgetsFlutterBinding.ensureInitialized())
             .setSurfaceSize(const Size(3e2, 4e3));
 
         await tester.pumpWidget(
